requirements.txt
Flask==2.0.1
gunicorn==20.1.0
requests==2.20.0
psycopg2==2.8.6
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Pyrebase4==4.5.0
# firebase-admin==5.0.0
psycopg2-pool==1.1